backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
aiosqlite>=0.19.0

# Environment and Settings
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP Client for M-Pesa Integration
httpx>=0.25.1
requests>=2.31.0

# Utility Packages
uuid>=1.30
python-dateutil>=2.8.2

# Optional - Notion Integration
notion-client>=2.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1

# Development Tools
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0